Reorganized BrokenDevice
Unfortunately working with Program Fixpoints/well-founded recursion is
harder than I anticipated, mostly because I don't know how to prove
anything with them. So I have siloed that stuff off to EncodeNat.v, and
moved the core definitions to Definitions.v. This hopefully makes it a
lot more clear which parts are not done (encoding nat) and which are
(the rest of the stuff).